split pkg/utils/util.go
