+++ pkg/utils/util.go
@@ -12,7 +12,6 @@ import (
 	"io"
 	"net"
 	"net/http"
-	"net/url"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -22,41 +21,18 @@ import (
 	"time"
 
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
-	"knative.dev/pkg/apis"
 
-	devfilePkg "github.com/devfile/library/v2/pkg/devfile"
-	"github.com/devfile/library/v2/pkg/devfile/parser"
-	"github.com/devfile/library/v2/pkg/devfile/parser/data"
 	"github.com/devfile/library/v2/pkg/util"
-	"github.com/google/go-containerregistry/pkg/authn"
-	"github.com/google/go-containerregistry/pkg/name"
-	remoteimg "github.com/google/go-containerregistry/pkg/v1/remote"
 	"github.com/mitchellh/go-homedir"
-	buildservice "github.com/redhat-appstudio/build-service/api/v1alpha1"
 	"github.com/redhat-appstudio/e2e-tests/pkg/constants"
-	"github.com/tektoncd/cli/pkg/bundle"
-	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
-	"github.com/tektoncd/pipeline/pkg/remote/oci"
 	"k8s.io/klog/v2"
-
-	"sigs.k8s.io/yaml"
-
-	crclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
 )
-
-type FailedPipelineRunDetails struct {
-	FailedTaskRunName   string
-	PodName             string
-	FailedContainerName string
-}
 
 type Options struct {
 	ToolchainApiUrl string
@@ -113,47 +89,6 @@ func GetEnvOrFunc(key string, defaultFunc func() (string, error)) (string, error
 	return defaultFunc()
 }
 
-/*
-Right now DevFile status in HAS is a string:
-metadata:
-
-	attributes:
-		appModelRepository.url: https://github.com/redhat-appstudio-qe/pet-clinic-application-service-establish-danger
-		gitOpsRepository.url: https://github.com/redhat-appstudio-qe/pet-clinic-application-service-establish-danger
-	name: pet-clinic
-	schemaVersion: 2.1.0
-
-The ObtainGitUrlFromDevfile extract from the string the git url associated with a application
-*/
-func ObtainGitOpsRepositoryName(devfileStatus string) string {
-	appDevfile, err := ParseDevfileModel(devfileStatus)
-	if err != nil {
-		err = fmt.Errorf("error parsing devfile: %v", err)
-	}
-	// Get the devfile attributes from the parsed object
-	devfileAttributes := appDevfile.GetMetadata().Attributes
-	gitOpsRepository := devfileAttributes.GetString("gitOpsRepository.url", &err)
-	parseUrl, err := url.Parse(gitOpsRepository)
-	if err != nil {
-		err = fmt.Errorf("fatal: %v", err)
-	}
-	repoParsed := strings.Split(parseUrl.Path, "/")
-
-	return repoParsed[len(repoParsed)-1]
-}
-
-func ObtainGitOpsRepositoryUrl(devfileStatus string) string {
-	appDevfile, err := ParseDevfileModel(devfileStatus)
-	if err != nil {
-		err = fmt.Errorf("error parsing devfile: %v", err)
-	}
-	// Get the devfile attributes from the parsed object
-	devfileAttributes := appDevfile.GetMetadata().Attributes
-	gitOpsRepository := devfileAttributes.GetString("gitOpsRepository.url", &err)
-
-	return gitOpsRepository
-}
-
 func GetQuayIOOrganization() string {
 	return GetEnv(constants.QUAY_E2E_ORGANIZATION_ENV, "redhat-appstudio-qe")
 }
@@ -183,30 +118,6 @@ func GetOpenshiftToken() (token string, err error) {
 		return "", fmt.Errorf("error obtaining oc token %s", err.Error())
 	}
 	return strings.TrimSuffix(string(tokenBytes), "\n"), nil
-}
-
-func GetFailedPipelineRunDetails(c crclient.Client, pipelineRun *v1beta1.PipelineRun) (*FailedPipelineRunDetails, error) {
-	d := &FailedPipelineRunDetails{}
-	for _, chr := range pipelineRun.Status.PipelineRunStatusFields.ChildReferences {
-		taskRun := &v1beta1.TaskRun{}
-		taskRunKey := types.NamespacedName{Namespace: pipelineRun.Namespace, Name: chr.Name}
-		if err := c.Get(context.Background(), taskRunKey, taskRun); err != nil {
-			return nil, fmt.Errorf("failed to get details for PR %s: %+v", pipelineRun.GetName(), err)
-		}
-		for _, c := range taskRun.Status.Conditions {
-			if c.Reason == "Failed" {
-				d.FailedTaskRunName = taskRun.Name
-				d.PodName = taskRun.Status.PodName
-				for _, s := range taskRun.Status.TaskRunStatusFields.Steps {
-					if s.Terminated.Reason == "Error" {
-						d.FailedContainerName = s.ContainerName
-						return d, nil
-					}
-				}
-			}
-		}
-	}
-	return d, nil
 }
 
 func GetGeneratedNamespace(name string) string {
@@ -304,102 +215,6 @@ func CreateDockerConfigFile(base64EncodedString string) error {
 	return nil
 }
 
-// ExtractTektonObjectFromBundle extracts specified Tekton object from specified bundle reference
-func ExtractTektonObjectFromBundle(bundleRef, kind, name string) (runtime.Object, error) {
-	var obj runtime.Object
-	var err error
-
-	resolver := oci.NewResolver(bundleRef, authn.DefaultKeychain)
-	if obj, _, err = resolver.Get(context.Background(), kind, name); err != nil {
-		return nil, fmt.Errorf("failed to fetch the tekton object %s with name %s: %v", kind, name, err)
-	}
-	return obj, nil
-}
-
-// BuildAndPushTektonBundle builds a Tekton bundle from YAML and pushes to remote container registry
-func BuildAndPushTektonBundle(YamlContent []byte, ref name.Reference, remoteOption remoteimg.Option) error {
-	img, err := bundle.BuildTektonBundle([]string{string(YamlContent)}, os.Stdout)
-	if err != nil {
-		return fmt.Errorf("error when building a bundle %s: %v", ref.String(), err)
-	}
-
-	outDigest, err := bundle.Write(img, ref, remoteOption)
-	if err != nil {
-		return fmt.Errorf("error when pushing a bundle %s to a container image registry repo: %v", ref.String(), err)
-	}
-	klog.Infof("image digest for a new tekton bundle %s: %+v", ref.String(), outDigest)
-
-	return nil
-}
-
-// GetDefaultPipelineBundleRef gets the specific Tekton pipeline bundle reference from a Build pipeline selector
-// (in a YAML format) from a URL specified in the parameter
-func GetDefaultPipelineBundleRef(buildPipelineSelectorYamlURL, selectorName string) (string, error) {
-	res, err := http.Get(buildPipelineSelectorYamlURL)
-	if err != nil {
-		return "", fmt.Errorf("failed to get a build pipeline selector from url %s: %v", buildPipelineSelectorYamlURL, err)
-	}
-	defer res.Body.Close()
-	body, err := io.ReadAll(res.Body)
-	if err != nil {
-		return "", fmt.Errorf("failed to read the body response of a build pipeline selector: %v", err)
-	}
-	ps := &buildservice.BuildPipelineSelector{}
-	if err = yaml.Unmarshal(body, ps); err != nil {
-		return "", fmt.Errorf("failed to unmarshal build pipeline selector: %v", err)
-	}
-	for i := range ps.Spec.Selectors {
-		s := &ps.Spec.Selectors[i]
-		if s.Name == selectorName {
-			return GetBundleRef(&s.PipelineRef), nil
-		}
-	}
-
-	return "", fmt.Errorf("could not find %s pipeline bundle in build pipeline selector fetched from %s", selectorName, buildPipelineSelectorYamlURL)
-}
-
-// GetPipelineNameAndBundleRef returns the pipeline name and bundle reference from a pipelineRef
-// https://tekton.dev/docs/pipelines/pipelineruns/#tekton-bundles
-func GetPipelineNameAndBundleRef(pipelineRef *v1beta1.PipelineRef) (string, string) {
-	var name string
-	var bundleRef string
-
-	// Prefer the v1 style
-	if pipelineRef.Resolver != "" {
-		for _, param := range pipelineRef.Params {
-			switch param.Name {
-			case "name":
-				name = param.Value.StringVal
-			case "bundle":
-				bundleRef = param.Value.StringVal
-			}
-		}
-	} else {
-		// Support the v1beta1 style
-		name = pipelineRef.Name
-		bundleRef = pipelineRef.Bundle //nolint:all
-	}
-
-	return name, bundleRef
-}
-
-// GetBundleRef returns the bundle reference from a pipelineRef
-func GetBundleRef(pipelineRef *v1beta1.PipelineRef) string {
-	_, bundleRef := GetPipelineNameAndBundleRef(pipelineRef)
-	return bundleRef
-}
-
-// ParseDevfileModel calls the devfile library's parse and returns the devfile data
-func ParseDevfileModel(devfileModel string) (data.DevfileData, error) {
-	// Retrieve the devfile from the body of the resource
-	devfileBytes := []byte(devfileModel)
-	parserArgs := parser.ParserArgs{
-		Data: devfileBytes,
-	}
-	devfileObj, _, err := devfilePkg.ParseDevfileAndValidate(parserArgs)
-	return devfileObj.Data, err
-}
-
 func HostIsAccessible(host string) bool {
 	tc := &http.Transport{
 		TLSClientConfig: &tls.Config{
@@ -412,14 +227,6 @@ func HostIsAccessible(host string) bool {
 		return false
 	}
 	return true
-}
-
-func HasPipelineRunSucceeded(pr *v1beta1.PipelineRun) bool {
-	return pr.GetStatusCondition().GetCondition(apis.ConditionSucceeded).IsTrue()
-}
-
-func HasPipelineRunFailed(pr *v1beta1.PipelineRun) bool {
-	return pr.IsDone() && pr.GetStatusCondition().GetCondition(apis.ConditionSucceeded).IsFalse()
 }
 
 // Return a container logs from a given pod and namespace
@@ -516,7 +323,6 @@ func CheckFileExistsInDir(rootDir, filename string) (bool, error) {
 }
 
 func Untar(dst string, tarPath string) error {
-
 	tr, err := ReadTarFile(tarPath)
 	if err != nil {
 		return err
