+++ tests/remote-secret/component-annotation-image-pull-remote-secret.go
@@ -9,6 +9,7 @@ import (
 	appservice "github.com/redhat-appstudio/application-api/api/v1alpha1"
 	"github.com/redhat-appstudio/e2e-tests/pkg/framework"
 	"github.com/redhat-appstudio/e2e-tests/pkg/utils"
+	"github.com/redhat-appstudio/e2e-tests/pkg/utils/gitops"
 	rs "github.com/redhat-appstudio/remote-secret/api/v1beta1"
 	"k8s.io/apimachinery/pkg/api/meta"
 )
@@ -77,7 +78,7 @@ var _ = framework.RemoteSecretSuiteDescribe(Label("component-annotation-image-pu
 			}, 3*time.Minute, 100*time.Millisecond).Should(Not(BeEmpty()), fmt.Sprintf("timed out waiting for gitOps repository to be created for the %s application in %s namespace", applicationName, fw.UserNamespace))
 
 			Eventually(func() bool {
-				gitOpsRepository := utils.ObtainGitOpsRepositoryName(application.Status.Devfile)
+				gitOpsRepository := gitops.ObtainGitOpsRepositoryName(application.Status.Devfile)
 
 				return fw.AsKubeDeveloper.CommonController.Github.CheckIfRepositoryExist(gitOpsRepository)
 			}, 1*time.Minute, 1*time.Second).Should(BeTrue(), fmt.Sprintf("timed out waiting for HAS controller to create gitops repository for the %s application in %s namespace", applicationName, fw.UserNamespace))
